+++ acp/acp-spring-boot-starter-common/src/main/java/pers/acp/springboot/common/conf/ScheduleConfiguration.java
@@ -6,6 +6,8 @@ import java.util.HashMap;
 import java.util.Map;
 
 /**
+ * 定时调度配置
+ *
  * @author zhangbin by 2018-1-20 21:08
  * @since JDK1.8
  */