+++ cloud/helloworld/src/main/java/pers/acp/springcloud/server/helloworld/feign/WorldServer.java
@@ -1,6 +1,7 @@
 package pers.acp.springcloud.server.helloworld.feign;
 
 import org.springframework.cloud.openfeign.FeignClient;
+import org.springframework.stereotype.Component;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -10,6 +11,7 @@ import pers.acp.springcloud.server.helloworld.hystrix.WorldServerHystrix;
  * @author zhangbin by 2018-3-6 15:32
  * @since JDK1.8
  */
+@Component
 @FeignClient(value = "atomic-world", fallbackFactory = WorldServerHystrix.class)
 public interface WorldServer {