+++ cloud/oauth-server/src/main/java/pers/acp/springcloud/oauth/domain/SecurityClientDetailsService.java
@@ -20,6 +20,9 @@ public class SecurityClientDetailsService implements ClientDetailsService {
 
     private ClientDetailsService clientDetailsService = null;
 
+    /**
+     * 初始化客户端信息
+     */
     @PostConstruct
     public void init() {
         InMemoryClientDetailsServiceBuilder builder = new InMemoryClientDetailsServiceBuilder();