+++ test/testspringboot/src/main/java/pers/acp/test/application/conf/JpaPgConfig.java
@@ -1,9 +1,11 @@
 package pers.acp.test.application.conf;
 
+import com.zaxxer.hikari.HikariConfig;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.orm.jpa.HibernateSettings;
 import org.springframework.boot.autoconfigure.orm.jpa.JpaProperties;
 import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.boot.jdbc.DataSourceBuilder;
 import org.springframework.boot.orm.jpa.EntityManagerFactoryBuilder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -12,9 +14,9 @@ import org.springframework.orm.jpa.JpaTransactionManager;
 import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
 import org.springframework.transaction.PlatformTransactionManager;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
-import pers.acp.springboot.core.datasource.AcpDataSource;
 
 import javax.persistence.EntityManager;
+import javax.sql.DataSource;
 import java.util.Objects;
 
 /**
@@ -37,16 +39,22 @@ public class JpaPgConfig {
 
     @Bean
     @ConfigurationProperties(prefix = "spring.datasource.pg")
-    public AcpDataSource pgDataSource() {
-        return new AcpDataSource();
+    public DataSource pgDataSource() {
+        return DataSourceBuilder.create().build();
+    }
+
+    @Bean
+    @ConfigurationProperties(prefix = "spring.datasource.pg")
+    public HikariConfig pgConfig() {
+        return new HikariConfig();
     }
 
     @Bean(name = "entityManagerFactoryPg")
     public LocalContainerEntityManagerFactoryBean entityManagerFactoryPrimary(EntityManagerFactoryBuilder builder) {
-        AcpDataSource pgDS = pgDataSource();
-        jpaProperties.getProperties().put("hibernate.dialect", pgDS.getDialect());
-        return builder.dataSource(pgDS)
-                .packages(pgDS.getScanpackage().split(","))
+        HikariConfig hikariConfig = pgConfig();
+        jpaProperties.getProperties().put("hibernate.dialect", hikariConfig.getDataSourceProperties().getProperty("dialect"));
+        return builder.dataSource(pgDataSource())
+                .packages(hikariConfig.getDataSourceProperties().getProperty("scanpackage").split(","))
                 .persistenceUnit("persistenceUnitPg")
                 .properties(jpaProperties.getHibernateProperties(new HibernateSettings()))
                 .build();