+++ doc/version_history.md
@@ -1,4 +1,11 @@
 ## 版本更新记录
+##### 4.0.1
+    1. SpringBoot 升级至 2.0.2.RELEASE
+    2. 修改数据源配置，连接池改用 HikariCP，去除 tomcat 连接池依赖
+    3. 升级 spring-cloud 版本为 Finchley.RC2
+    4. 增加相关注释
+    5. 修复 acp-client 中 webservice client 调用异常的bug，增加 webservice client 用法 demo
+    6. oauth-server 中，token 持久化到 redis
 ##### 4.0.0
     1. 优化 gradle 脚本，spring cloud 版本号写入 dependencies.gradle；删除 cloud 模块下的 build.gradle
     2. 升级 spring-cloud 版本为 Finchley.RC1